Use uint8_t as data types for input and output of the ISP.
The ISP will accept/produce an uint8_t image array. Interally, it
works on half-point float data.

+++ src/camera_pipe.h
@@ -1,7 +1,7 @@
 #ifndef _CAMERA_PIPE_H_
 #define _CAMERA_PIPE_H_
 
-void camera_pipe(float *host_input, float *host_result, int row_size,
+void camera_pipe(uint8_t *host_input, uint8_t *host_result, int row_size,
                  int col_size);
 
 #endif
